refactor: migrate OpenUpwardsPageTransitionsBuilder to widgets
