+++ packages/flutter/lib/src/material/page_transitions_theme.dart
@@ -20,146 +20,6 @@ import 'colors.dart';
 import 'predictive_back_page_transitions_builder.dart';
 import 'theme.dart';
 
-// This transition is intended to match the default for Android P.
-class _OpenUpwardsPageTransition extends StatefulWidget {
-  const _OpenUpwardsPageTransition({
-    required this.animation,
-    required this.secondaryAnimation,
-    required this.child,
-  });
-
-  // The new page slides upwards just a little as its clip
-  // rectangle exposes the page from bottom to top.
-  static final Tween<Offset> _primaryTranslationTween = Tween<Offset>(
-    begin: const Offset(0.0, 0.05),
-    end: Offset.zero,
-  );
-
-  // The old page slides upwards a little as the new page appears.
-  static final Tween<Offset> _secondaryTranslationTween = Tween<Offset>(
-    begin: Offset.zero,
-    end: const Offset(0.0, -0.025),
-  );
-
-  // The scrim obscures the old page by becoming increasingly opaque.
-  static final Tween<double> _scrimOpacityTween = Tween<double>(begin: 0.0, end: 0.25);
-
-  // Used by all of the transition animations.
-  static const Curve _transitionCurve = Cubic(0.20, 0.00, 0.00, 1.00);
-
-  final Animation<double> animation;
-  final Animation<double> secondaryAnimation;
-  final Widget child;
-
-  @override
-  State<_OpenUpwardsPageTransition> createState() => _OpenUpwardsPageTransitionState();
-}
-
-class _OpenUpwardsPageTransitionState extends State<_OpenUpwardsPageTransition> {
-  late CurvedAnimation _primaryAnimation;
-  late CurvedAnimation _secondaryTranslationCurvedAnimation;
-
-  @override
-  void initState() {
-    super.initState();
-    _setAnimations();
-  }
-
-  @override
-  void didUpdateWidget(covariant _OpenUpwardsPageTransition oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.animation != widget.animation ||
-        oldWidget.secondaryAnimation != widget.secondaryAnimation) {
-      _disposeAnimations();
-      _setAnimations();
-    }
-  }
-
-  void _setAnimations() {
-    _primaryAnimation = CurvedAnimation(
-      parent: widget.animation,
-      curve: _OpenUpwardsPageTransition._transitionCurve,
-      reverseCurve: _OpenUpwardsPageTransition._transitionCurve.flipped,
-    );
-    _secondaryTranslationCurvedAnimation = CurvedAnimation(
-      parent: widget.secondaryAnimation,
-      curve: _OpenUpwardsPageTransition._transitionCurve,
-      reverseCurve: _OpenUpwardsPageTransition._transitionCurve.flipped,
-    );
-  }
-
-  void _disposeAnimations() {
-    _primaryAnimation.dispose();
-    _secondaryTranslationCurvedAnimation.dispose();
-  }
-
-  @override
-  void dispose() {
-    _disposeAnimations();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return LayoutBuilder(
-      builder: (BuildContext context, BoxConstraints constraints) {
-        final Size size = constraints.biggest;
-
-        // Gradually expose the new page from bottom to top.
-        final Animation<double> clipAnimation = Tween<double>(
-          begin: 0.0,
-          end: size.height,
-        ).animate(_primaryAnimation);
-
-        final Animation<double> opacityAnimation = _OpenUpwardsPageTransition._scrimOpacityTween
-            .animate(_primaryAnimation);
-        final Animation<Offset> primaryTranslationAnimation = _OpenUpwardsPageTransition
-            ._primaryTranslationTween
-            .animate(_primaryAnimation);
-
-        final Animation<Offset> secondaryTranslationAnimation = _OpenUpwardsPageTransition
-            ._secondaryTranslationTween
-            .animate(_secondaryTranslationCurvedAnimation);
-
-        return AnimatedBuilder(
-          animation: widget.animation,
-          builder: (BuildContext context, Widget? child) {
-            return ColoredBox(
-              color: Colors.black.withOpacity(opacityAnimation.value),
-              child: Align(
-                alignment: Alignment.bottomLeft,
-                child: ClipRect(
-                  child: SizedBox(
-                    height: clipAnimation.value,
-                    child: OverflowBox(
-                      alignment: Alignment.bottomLeft,
-                      maxHeight: size.height,
-                      child: child,
-                    ),
-                  ),
-                ),
-              ),
-            );
-          },
-          child: AnimatedBuilder(
-            animation: widget.secondaryAnimation,
-            child: FractionalTranslation(
-              translation: primaryTranslationAnimation.value,
-              child: widget.child,
-            ),
-            builder: (BuildContext context, Widget? child) {
-              return FractionalTranslation(
-                translation: secondaryTranslationAnimation.value,
-                child: child,
-              );
-            },
-          ),
-        );
-      },
-    );
-  }
-}
-
 // Zooms and fades a new page in, zooming out the previous page. This transition
 // is designed to match the Android Q activity transition.
 class _ZoomPageTransition extends StatelessWidget {
@@ -560,43 +420,6 @@ class _FadeForwardsPageTransition extends StatelessWidget {
         backgroundColor,
         child,
       ),
-    );
-  }
-}
-
-/// Used by [PageTransitionsTheme] to define a vertical [MaterialPageRoute] page
-/// transition animation that looks like the default page transition
-/// used on Android P.
-///
-/// See also:
-///
-///  * [FadeUpwardsPageTransitionsBuilder], which defines a page transition
-///    that's similar to the one provided by Android O.
-///  * [ZoomPageTransitionsBuilder], which defines the default page transition
-///    that's similar to the one provided in Android Q.
-///  * [CupertinoPageTransitionsBuilder], which defines a horizontal page
-///    transition that matches native iOS page transitions.
-///  * [PredictiveBackPageTransitionsBuilder], which defines a page
-///    transition that allows peeking behind the current route on Android.
-///  * [FadeForwardsPageTransitionsBuilder], which defines a page transition
-///    that's similar to the one provided by Android U.
-class OpenUpwardsPageTransitionsBuilder extends PageTransitionsBuilder {
-  /// Constructs a page transition animation that matches the transition used on
-  /// Android P.
-  const OpenUpwardsPageTransitionsBuilder();
-
-  @override
-  Widget buildTransitions<T>(
-    PageRoute<T>? route,
-    BuildContext? context,
-    Animation<double> animation,
-    Animation<double> secondaryAnimation,
-    Widget child,
-  ) {
-    return _OpenUpwardsPageTransition(
-      animation: animation,
-      secondaryAnimation: secondaryAnimation,
-      child: child,
     );
   }
 }
