+++ packages/flutter/test/widgets/page_transitions_builder_test.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -317,6 +318,188 @@ void main() {
 
       await tester.pump(const Duration(milliseconds: 300));
 
+      expect(find.text('Page 1'), isOnstage);
+      expect(find.text('Page 2'), findsNothing);
+    },
+    variant: TargetPlatformVariant.only(TargetPlatform.android),
+  );
+
+  testWidgets(
+    'PageTransitionsTheme override builds a _OpenUpwardsPageTransition',
+    (WidgetTester tester) async {
+      final Map<String, WidgetBuilder> routes = <String, WidgetBuilder>{
+        '/': (BuildContext context) => Material(
+          child: TextButton(
+            child: const Text('push'),
+            onPressed: () {
+              Navigator.of(context).pushNamed('/b');
+            },
+          ),
+        ),
+        '/b': (BuildContext context) => const Text('page b'),
+      };
+
+      await tester.pumpWidget(
+        MaterialApp(
+          theme: ThemeData(
+            pageTransitionsTheme: const PageTransitionsTheme(
+              builders: <TargetPlatform, PageTransitionsBuilder>{
+                TargetPlatform.android:
+                    OpenUpwardsPageTransitionsBuilder(), // creates a _OpenUpwardsPageTransition
+              },
+            ),
+          ),
+          routes: routes,
+        ),
+      );
+
+      Finder findOpenUpwardsPageTransition() {
+        return find.descendant(
+          of: find.byType(MaterialApp),
+          matching: find.byWidgetPredicate(
+            (Widget w) => '${w.runtimeType}' == '_OpenUpwardsPageTransition',
+          ),
+        );
+      }
+
+      expect(
+        Theme.of(tester.element(find.text('push'))).platform,
+        debugDefaultTargetPlatformOverride,
+      );
+      expect(findOpenUpwardsPageTransition(), findsOneWidget);
+
+      await tester.tap(find.text('push'));
+      await tester.pumpAndSettle();
+      expect(find.text('page b'), findsOneWidget);
+      expect(findOpenUpwardsPageTransition(), findsOneWidget);
+    },
+    variant: TargetPlatformVariant.only(TargetPlatform.android),
+  );
+
+  testWidgets('OpenUpwardsPageTransitionsBuilder test', (WidgetTester tester) async {
+    await tester.pumpWidget(
+      Directionality(
+        textDirection: TextDirection.ltr,
+        child: Navigator(
+          onGenerateRoute: (RouteSettings settings) {
+            return _CustomPageRoute<void>(
+              settings: settings,
+              transitionsBuilder: const OpenUpwardsPageTransitionsBuilder(),
+              builder: (BuildContext context) {
+                if (settings.name == '/') {
+                  return ColoredBox(
+                    color: const Color(0xFF2196F3),
+                    child: Center(
+                      child: GestureDetector(
+                        onTap: () {
+                          Navigator.of(context).pushNamed('/second');
+                        },
+                        child: const Text(
+                          'Page 1',
+                          style: TextStyle(color: Color(0xFFFFFFFF), fontSize: 24),
+                        ),
+                      ),
+                    ),
+                  );
+                }
+                return const ColoredBox(
+                  color: Color(0xFF4CAF50),
+                  child: Center(
+                    child: Text('Page 2', style: TextStyle(color: Color(0xFFFFFFFF), fontSize: 24)),
+                  ),
+                );
+              },
+            );
+          },
+        ),
+      ),
+    );
+
+    final Offset widget1TopLeft = tester.getTopLeft(find.text('Page 1'));
+
+    await tester.tap(find.text('Page 1'));
+    await tester.pump();
+    await tester.pump(const Duration(milliseconds: 1));
+
+    expect(find.text('Page 1'), findsOneWidget);
+    expect(find.text('Page 2'), findsOneWidget);
+
+    final Offset widget2TopLeft = tester.getTopLeft(find.text('Page 2'));
+
+    expect(widget1TopLeft.dx, widget2TopLeft.dx);
+    expect(widget1TopLeft.dy <= widget2TopLeft.dy, true);
+
+    await tester.pump(const Duration(milliseconds: 300));
+
+    // After animation, only Page 2 should be visible.
+    expect(find.text('Page 1'), findsNothing);
+    expect(find.text('Page 2'), isOnstage);
+
+    tester.state<NavigatorState>(find.byType(Navigator)).pop();
+    await tester.pump();
+    await tester.pump(const Duration(milliseconds: 1));
+
+    expect(find.text('Page 1'), findsOneWidget);
+    expect(find.text('Page 2'), findsOneWidget);
+
+    await tester.pump(const Duration(milliseconds: 300));
+
+    // After reverse animation, only Page 1 should be visible.
+    expect(find.text('Page 1'), isOnstage);
+    expect(find.text('Page 2'), findsNothing);
+  });
+
+  testWidgets(
+    'OpenUpwardsPageTransitionsBuilder test with Material PageTransitionTheme',
+    (WidgetTester tester) async {
+      await tester.pumpWidget(
+        MaterialApp(
+          home: const Material(child: Text('Page 1')),
+          theme: ThemeData(
+            pageTransitionsTheme: const PageTransitionsTheme(
+              builders: <TargetPlatform, PageTransitionsBuilder>{
+                TargetPlatform.android: OpenUpwardsPageTransitionsBuilder(),
+              },
+            ),
+          ),
+          routes: <String, WidgetBuilder>{
+            '/next': (BuildContext context) {
+              return const Material(child: Text('Page 2'));
+            },
+          },
+        ),
+      );
+
+      final Offset widget1TopLeft = tester.getTopLeft(find.text('Page 1'));
+
+      tester.state<NavigatorState>(find.byType(Navigator)).pushNamed('/next');
+      await tester.pump();
+      await tester.pump(const Duration(milliseconds: 1));
+
+      expect(find.text('Page 1'), findsOneWidget);
+      expect(find.text('Page 2'), findsOneWidget);
+
+      final Offset widget2TopLeft = tester.getTopLeft(find.text('Page 2'));
+
+      expect(widget1TopLeft.dx, widget2TopLeft.dx);
+      expect(widget1TopLeft.dy < widget2TopLeft.dy, true);
+
+      await tester.pump(const Duration(milliseconds: 300));
+
+      // Page 2 covers page 1.
+      expect(find.text('Page 1'), findsNothing);
+      expect(find.text('Page 2'), isOnstage);
+
+      tester.state<NavigatorState>(find.byType(Navigator)).pop();
+      await tester.pump();
+      await tester.pump(const Duration(milliseconds: 1));
+
+      expect(find.text('Page 1'), findsOneWidget);
+      expect(find.text('Page 2'), findsOneWidget);
+
+      await tester.pump(const Duration(milliseconds: 300));
+
+      // Back to page 1.
       expect(find.text('Page 1'), isOnstage);
       expect(find.text('Page 2'), findsNothing);
     },
